Remove scopes of routes unregistered with the server.table() and/or without an auth strategy. Cleaned up prints and console.log()
