+++ server/api/index.js
@@ -2,7 +2,7 @@
 
 exports.register = function (server, options, next) {
 
-  server.route({//TODO: this route didn't have any strategies or auth before
+  server.route({
     method: 'GET',
     path: '/',
     handler: function (request, reply) {
