+++ server/api/users.js
@@ -10,7 +10,6 @@ const DEFAULT_ROLES = require('../helper/getDefaultRoles');
 const internals = {};
 
 
-
 internals.applyRoutes = function (server, next) {
 
   const User = server.plugins['hicsail-hapi-mongo-models'].User;
@@ -905,6 +904,26 @@ internals.applyRoutes = function (server, next) {
     },
     handler: function (request, reply) {
 
+      const x = server.table()[0].table;
+      x.forEach((item) => {
+
+        if (item.hasOwnProperty('path')){//processing routes in server
+          if (item.path.includes('templates')){
+            console.log('templates');
+          }
+          const path = item.path;
+          const method = item.method;
+          if (item.settings.hasOwnProperty('auth') && typeof item.settings.auth !== 'undefined' && item.settings.auth.hasOwnProperty('access') ){
+            // console.log(item.settings.auth.access[0].scope.selection);
+          }
+          else {//routes don't have scope, assign default value to each route [‘root’, ‘admin’ ,’researcher’, ‘analyst’,’ clinician‘]
+            // console.log('[‘root’, ‘admin’ ,’researcher’, ‘analyst’,’ clinician‘]');
+          }
+        }
+      });
+      // PermissionConfig.createTable(server, (result) => {
+      //   console.log(result);
+      // });
       const pathScopeReference = PermissionConfigTable[request.payload.method][request.payload.path];
       if (pathScopeReference.includes(request.payload.role)){
         pathScopeReference.splice(pathScopeReference.indexOf(request.payload.role), 1);
