--- server/web/helpers/routeMethod.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = (route) => {//a helper used to format the object differently
-
-  return route.substring(0, route.search('/'));
-};
