--- server/web/helpers/routePath.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = (route) => {//a helper used to format the object differently
-
-  console.log(route);
-  // return route.substring(route.search('/'));
-};
