+++ server/web/routes/public.js
@@ -1,10 +1,8 @@
 'use strict';
-const PermissionConfigTable = require('../../../permission-config');
-const DEFAULT_ROLES = require('../../helper/getDefaultRoles');
 
 exports.register = function (server, options, next) {
 
-  server.route({//TODO: this route didn't have any strategies or auth before AND Error occurs when new scope is applied
+  server.route({
     method: 'GET',
     path: '/public/{param*}',
     config: {
